// src/app/models/tag.ts
export interface Tag {
  id: number;
  name: string;
  archival?: boolean;
}


//  class TagClass implements Tag {
//   constructor(private _id: number, private _name: string) {
//   }
//
//   get id() {
//     return this._id
//   }
//
//   set id(value) {
//     this._id = value
//   }
//
//   get name() {
//     return this._name
//   }
//
//   set name(value) {
//     console.log('new name', value)
//     this._name = value
//   }
// }